+++ src/lib/incident.ts
@@ -2,6 +2,18 @@ import objectInspect from "object-inspect";
 import { Incident as Interface, StaticIncident as StaticInterface } from "./types";
 
 /**
+ * Default message formatter.
+ *
+ * This uses `object-inspect` to print the `data` object.
+ *
+ * @param data Data object associated with the error.
+ * @return String representation of the data object.
+ */
+export function format(data: any): string {
+  return objectInspect(data, {depth: 30});
+}
+
+/**
  * Define a hidden property.
  *
  * @param obj
@@ -148,7 +160,7 @@ function createIncident(_super: Function): StaticInterface {
       message = args[argIndex];
     } else {
       if (data !== undefined) {
-        message = objectInspect;
+        message = format;
       } else {
         message = "";
       }